// src/user/user.type.ts
export type UserType = {
  /** (строка): Имя пользователя или уникальный идентификатор аккаунта. */
  username: string;

  /** (строка | null): Полное имя пользователя. */
  full_name?: string;

  /** (строка): E-mail пользователя. */
  email?: string;

  /** (строка): Пароль пользователя. */
  password?: string;

  /** (строка): Номер телефона пользователя. */
  phone?: string;

  /** (строка | null): Биография или описание пользователя. */
  bio?: string;

  /** (строка | null): Ссылка на веб-сайт пользователя. */
  website?: string;

  /** (строка | null): URL профильной фотографии пользователя. */
  profile_picture?: string;

  /** (булево значение): Флаг, указывающий, является ли профиль пользователя приватным. */
  is_private: boolean;

  /** (булево значение): Флаг, указывающий, подтвержден ли аккаунт пользователя. */
  is_verified: boolean;
};